更新gorm配置

+++ xapi-server/application/controllers/MessageController.go
@@ -186,6 +186,6 @@ func GetUnreadMessage(c *gin.Context) {
 	c.JSON(http.StatusOK, gin.H{
 		"status":  200,
 		"message": "成功",
-		"data":    map[string]int{"count": count},
+		"data":    map[string]int64{"count": count},
 	})
 }
